move message history out to utils
