+++ environments/game_environments/gymnasium/blackjack_env.py
@@ -27,8 +27,12 @@ from atroposlib.envs.base import (
     OpenaiConfig,
     ScoredDataGroup,
 )
-from atroposlib.utils.tokenize_for_trainer import tokenize_for_trainer
-from atroposlib.utils.tool_call_parser import parse_tool_call
+from atroposlib.utils import (
+    tokenize_for_trainer,
+    parse_tool_call,
+    truncate_thinking,
+    ensure_trajectory_token_limit
+)
 
 logger = logging.getLogger(__name__)
 
@@ -42,7 +46,7 @@ class BlackjackEnvConfig(BaseEnvConfig):
     thinking_active: bool = True
     eval_episodes: int = 100
     max_think_chars_history: int = 3000
-    max_trajectory_tokens: int = 24576
+    max_trajectory_tokens: int = 24576 #seq_len of RL trainer
     debug_mode: bool = False
     group_size: int = 16
     tiebreak_token_factor: float = 0.01
@@ -558,7 +562,7 @@ class BlackjackEnv(BaseEnv):
 
             ep.message_history = current_state_messages
 
-            response_for_history = self._truncate_thinking_for_history(
+            response_for_history = truncate_thinking(
                 chosen_full_response, self.config.max_think_chars_history
             )
             ep.message_history.append(
@@ -649,7 +653,11 @@ class BlackjackEnv(BaseEnv):
                 )
             del self.episodes[seed]
 
-        return self._ensure_trajectory_token_limit(trajectory_data_for_trainer)
+        return ensure_trajectory_token_limit(
+            trajectory_data_for_trainer,
+            self.tokenizer,
+            self.config.max_trajectory_tokens,
+        )
 
     async def score(
         self, rollout_group_data: List[BlackjackScoredDataGroup]
@@ -721,7 +729,7 @@ class BlackjackEnv(BaseEnv):
             metrics["total_reward"] += reward
             metrics["num_turns"] = turn + 1
 
-            response_for_history = self._truncate_thinking_for_history(
+            response_for_history = truncate_thinking(
                 full_agent_response, self.config.max_think_chars_history
             )
 
@@ -860,271 +868,6 @@ class BlackjackEnv(BaseEnv):
             )
         ]
         return env_config, server_configs
-
-    def _truncate_thinking_for_history(self, response_text: str, max_chars: int) -> str:
-        """Helper to truncate the <think> block of a response for message history."""
-        try:
-            think_start_tag = "<think>"
-            think_end_tag = "</think>"
-
-            think_start_idx = response_text.find(think_start_tag)
-            think_end_idx = response_text.find(think_end_tag)
-
-            if (
-                think_start_idx != -1
-                and think_end_idx != -1
-                and think_start_idx < think_end_idx
-            ):
-                part_before_content = response_text[
-                    : think_start_idx + len(think_start_tag)
-                ]
-                original_think_content = response_text[
-                    think_start_idx + len(think_start_tag) : think_end_idx
-                ].strip()
-                part_after_content = response_text[think_end_idx:]
-
-                truncated_think_content = original_think_content
-                is_truncated = False
-
-                if not original_think_content:
-                    return response_text
-
-                paragraphs = [
-                    p.strip() for p in original_think_content.split("\n\n") if p.strip()
-                ]
-                if len(paragraphs) > 0:
-                    last_paragraph = paragraphs[-1]
-                    if len(last_paragraph) < len(original_think_content):
-                        truncated_think_content = last_paragraph
-                        is_truncated = True
-                    elif len(original_think_content) > max_chars:
-                        truncated_think_content = original_think_content[-max_chars:]
-                        is_truncated = True
-                elif len(original_think_content) > max_chars:
-                    truncated_think_content = original_think_content[-max_chars:]
-                    is_truncated = True
-
-                if is_truncated and truncated_think_content:
-                    if not truncated_think_content.startswith("... "):
-                        truncated_think_content = (
-                            "... " + truncated_think_content.lstrip()
-                        )
-
-                if (
-                    not truncated_think_content.strip()
-                    or truncated_think_content.strip() == "..."
-                ):
-                    final_content_for_block = ""
-                else:
-                    final_content_for_block = f"\n{truncated_think_content.strip()}\n"
-
-                return f"{part_before_content.rstrip()}{final_content_for_block}{part_after_content.lstrip()}"
-
-            return response_text
-        except Exception as e:
-            logger.error(
-                f"Error in _truncate_thinking_for_history for text '{response_text[:200]}...': {e}",
-                exc_info=True,
-            )
-            return response_text
-
-    def _ensure_trajectory_token_limit(
-        self, trajectory: List[BlackjackScoredDataGroup]
-    ) -> List[BlackjackScoredDataGroup]:
-        """
-        Ensure token sequences in a trajectory don't exceed max_trajectory_tokens.
-        Attempts to uniformly truncate older messages (preferably paired turns) from all alternatives within a step.
-        The system prompt, last environment observation, and last agent response are preserved as a minimum.
-        If a step still exceeds the limit after maximum possible truncation, it is discarded.
-
-        Args:
-            trajectory: List of BlackjackScoredDataGroup from an episode
-
-        Returns:
-            The trajectory with potentially truncated messages/tokens/masks or filtered steps
-        """
-        if not trajectory:
-            return trajectory
-
-        filtered_trajectory: List[BlackjackScoredDataGroup] = []
-
-        for step_idx, original_step_data in enumerate(trajectory):
-            if not (
-                original_step_data.get("messages")
-                and original_step_data.get("tokens")
-                and original_step_data.get("masks")
-                and original_step_data.get("seed") is not None
-                and original_step_data.get("parsed_actions") is not None
-            ):
-                logger.warning(
-                    f"[_ensure_trajectory_token_limit] Step {step_idx} in MC env "
-                    f"is missing critical data. Skipping."
-                )
-                continue
-
-            max_initial_tokens = 0
-            if original_step_data["tokens"]:
-                max_initial_tokens = (
-                    max(
-                        len(alt_tokens)
-                        for alt_tokens in original_step_data["tokens"]
-                        if isinstance(alt_tokens, list)
-                    )
-                    if any(
-                        isinstance(alt_tokens, list)
-                        for alt_tokens in original_step_data["tokens"]
-                    )
-                    else 0
-                )
-
-            if max_initial_tokens <= self.config.max_trajectory_tokens:
-                filtered_trajectory.append(original_step_data)
-                logger.info(
-                    f"[_ensure_trajectory_token_limit] Step {step_idx} compliant in MC env. "
-                    f"Max tokens: {max_initial_tokens}"
-                )
-                continue
-
-            logger.info(
-                f"[_ensure_trajectory_token_limit] Step {step_idx} in MC env (max tokens: {max_initial_tokens}) "
-                f"exceeds limit ({self.config.max_trajectory_tokens}). Attempting truncation."
-            )
-
-            working_messages = [
-                msgs_list.copy() for msgs_list in original_step_data["messages"] or []
-            ]
-            working_tokens = [
-                tkns_list.copy() for tkns_list in original_step_data["tokens"] or []
-            ]
-            working_masks = [
-                msks_list.copy() for msks_list in original_step_data["masks"] or []
-            ]
-            max_current_tokens = max_initial_tokens
-            num_alternatives = len(working_messages)
-
-            if num_alternatives == 0:
-                logger.warning(
-                    f"[_ensure_trajectory_token_limit] Step {step_idx} in MC env has no alternatives"
-                    " after copying. Skipping."
-                )
-                continue
-
-            retokenization_error_this_step = False
-            while max_current_tokens > self.config.max_trajectory_tokens:
-                target_pop_counts_per_alt = []
-                for alt_idx in range(num_alternatives):
-                    alt_msg_list = working_messages[alt_idx]
-                    num_preserved_at_end = 0
-                    if len(alt_msg_list) > 1 and alt_msg_list[-1]["role"] in [
-                        "agent",
-                        "assistant",
-                    ]:
-                        num_preserved_at_end = 1
-                        if (
-                            len(alt_msg_list) > 2
-                            and alt_msg_list[-2]["role"] == "environment"
-                        ):
-                            num_preserved_at_end = 2
-
-                    available_to_pop = len(alt_msg_list) - 1 - num_preserved_at_end
-
-                    if available_to_pop <= 0:
-                        target_pop_counts_per_alt.append(0)
-                    else:
-                        can_pop_pair = (
-                            available_to_pop >= 2
-                            and len(alt_msg_list) > 2
-                            and alt_msg_list[1]["role"] == "environment"
-                            and alt_msg_list[2]["role"] in ["agent", "assistant"]
-                        )
-                        if can_pop_pair:
-                            target_pop_counts_per_alt.append(2)
-                        else:
-                            target_pop_counts_per_alt.append(1)
-
-                positive_pop_counts = [c for c in target_pop_counts_per_alt if c > 0]
-                if not positive_pop_counts:
-                    break
-
-                min_pop_this_round = min(positive_pop_counts)
-                temp_new_alt_tokens = []
-                temp_new_alt_masks = []
-                max_tokens_after_this_trunc = 0
-
-                for alt_idx in range(num_alternatives):
-                    for _ in range(min_pop_this_round):
-                        if len(working_messages[alt_idx]) > 1:
-                            working_messages[alt_idx].pop(1)
-                        else:
-                            logger.error(
-                                f"[_ensure_trajectory_token_limit] MC env: Critical error during pop for "
-                                f"alt {alt_idx}, step {step_idx}. List too short."
-                            )
-                            retokenization_error_this_step = True
-                            break
-                    if retokenization_error_this_step:
-                        break
-
-                    try:
-                        tokenized_alt = tokenize_for_trainer(
-                            self.tokenizer, working_messages[alt_idx]
-                        )
-                        temp_new_alt_tokens.append(tokenized_alt["tokens"])
-                        temp_new_alt_masks.append(tokenized_alt["masks"])
-                        max_tokens_after_this_trunc = max(
-                            max_tokens_after_this_trunc, len(tokenized_alt["tokens"])
-                        )
-                    except Exception as e:
-                        logger.error(
-                            f"[_ensure_trajectory_token_limit] MC env: Error re-tokenizing alt {alt_idx} "
-                            f"in step {step_idx} after truncation: {e}"
-                        )
-                        retokenization_error_this_step = True
-                        break
-
-                if retokenization_error_this_step:
-                    break
-
-                working_tokens = temp_new_alt_tokens
-                working_masks = temp_new_alt_masks
-                max_current_tokens = max_tokens_after_this_trunc
-                logger.debug(
-                    f"[_ensure_trajectory_token_limit] MC env: Step {step_idx}, "
-                    f"after uniform pop of {min_pop_this_round}, "
-                    f"max tokens: {max_current_tokens}"
-                )
-
-            if (
-                not retokenization_error_this_step
-                and max_current_tokens <= self.config.max_trajectory_tokens
-            ):
-                updated_step_data: BlackjackScoredDataGroup = {
-                    "seed": original_step_data["seed"],
-                    "messages": working_messages,
-                    "tokens": working_tokens,
-                    "masks": working_masks,
-                    "scores": original_step_data.get("scores"),
-                    "parsed_actions": original_step_data.get("parsed_actions"),
-                }
-                filtered_trajectory.append(updated_step_data)
-                logger.info(
-                    f"[_ensure_trajectory_token_limit] MC env: Step {step_idx} successfully processed. "
-                    f"Final max tokens: {max_current_tokens}"
-                )
-            else:
-                logger.warning(
-                    f"[_ensure_trajectory_token_limit] MC env: Discarding step {step_idx}. "
-                    f"Max tokens ({max_current_tokens}) still exceed limit ({self.config.max_trajectory_tokens}) "
-                    f"or retokenization error occurred ({retokenization_error_this_step})."
-                )
-
-        if len(filtered_trajectory) < len(trajectory):
-            logger.warning(
-                f"[_ensure_trajectory_token_limit] MC env: Filtered out "
-                f"{len(trajectory) - len(filtered_trajectory)} steps "
-                f"due to token limit constraints. Original: {len(trajectory)}, Filtered: {len(filtered_trajectory)}"
-            )
-        return filtered_trajectory
 
     @classmethod
     def cli(cls):
